Criando a tabela de relatorios. Adicionando confirmação na exclusão de dados das tabelas.

+++ Sistema-de-gestão-de-controle/index.php
@@ -96,6 +96,11 @@ $usuarioLogado = $usuarioDao->autenticacao($_SESSION['token']);
     $lista = $vendaDao->getAll();
     ?>
 
+    <?php
+        $vendaDao = new daoVendas($conexao);
+        $relatorio = $vendaDao->getRelatorio();
+    ?>
+
     <div class="container">
         <div id="tb-vendas">
             <h1>Tabela de vendas</h1>
@@ -131,7 +136,7 @@ $usuarioLogado = $usuarioDao->autenticacao($_SESSION['token']);
                             </a>
                         </td>
                         <td>
-                            <a style="background-color:blue; border-radius:20px; padding:5px;" href="ExcluirVendas.php?id=<?php echo $item->getId() ?>">
+                            <a onClick="return confirm('Você tem certeza?');" style="background-color:blue; border-radius:20px; padding:5px;" href="ExcluirVendas.php?id=<?php echo $item->getId() ?>">
                                 <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="white" class="bi bi-trash3-fill" viewBox="0 0 16 16">
                                     <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z" />
                                 </svg>
@@ -173,7 +178,7 @@ $usuarioLogado = $usuarioDao->autenticacao($_SESSION['token']);
                             </a>
                         </td>
                         <td>
-                            <a style="background-color:blue; border-radius:20px; padding:5px;" href="ExcluirCompras.php?id=<?php echo $item->getId() ?>">
+                            <a onClick="return confirm('Você tem certeza?');" style="background-color:blue; border-radius:20px; padding:5px;" href="ExcluirCompras.php?id=<?php echo $item->getId() ?>">
                                 <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="white" class="bi bi-trash3-fill" viewBox="0 0 16 16">
                                     <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z" />
                                 </svg>
@@ -219,7 +224,7 @@ $usuarioLogado = $usuarioDao->autenticacao($_SESSION['token']);
                         </td>
                         <td>
                             <a style="background-color:blue; border-radius:20px; padding:5px;" href="ExcluirUsuarios.php?id=<?php echo $item->getId() ?>">
-                                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="white" class="bi bi-trash3-fill" viewBox="0 0 16 16">
+                                <svg onClick="return confirm('Você tem certeza?');" xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="white" class="bi bi-trash3-fill" viewBox="0 0 16 16">
                                     <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z" />
                                 </svg>
                             </a>
@@ -261,7 +266,7 @@ $usuarioLogado = $usuarioDao->autenticacao($_SESSION['token']);
                         </td>
                         <td>
                             <a style="background-color:blue; border-radius:20px; padding:5px;" href="ExcluirProdutos.php?id=<?php echo $item->getId() ?>">
-                                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="white" class="bi bi-trash3-fill" viewBox="0 0 16 16">
+                                <svg onClick="return confirm('Você tem certeza?');" xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="white" class="bi bi-trash3-fill" viewBox="0 0 16 16">
                                     <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z" />
                                 </svg>
                             </a>
@@ -272,34 +277,25 @@ $usuarioLogado = $usuarioDao->autenticacao($_SESSION['token']);
         </div>
 
         <!-- Tabela de relatorios -->
+        
+
         <div id="tb-relatorios">
             <h1>Tabela de Relatorios</h1>
             <table>
                 <tr>
-                    <th>Quantidade</th>
                     <th>Nome Vendedor</th>
+                    <th>Vendas</th>
                     <th>Valor</th>
-                    <th colspan="2">Ações</th>
                 </tr>
+                
+                <?php foreach ($relatorio as $item) : ?>
                 <tr>
-                    <td></td>
-                    <td></td>
-                    <td></td>
-                    <td>
-                        <a style="background-color:yellow; border-radius:20px; padding:5px;">
-                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="black" class="bi bi-pencil" viewBox="0 0 16 16">
-                                <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0 1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z" />
-                            </svg>
-                        </a>
-                    </td>
-                    <td>
-                        <a style="background-color:blue; border-radius:20px; padding:5px;">
-                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="white" class="bi bi-trash3-fill" viewBox="0 0 16 16">
-                                <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z" />
-                            </svg>
-                        </a>
-                    </td>
+                    <td><?=$item["seller"]?></td>
+                    <td><?=$item["qtd"]?></td>
+                    <td><?=round($item["total"], 2)?></td><!--Maskara para valores [Round]-->
                 </tr>
+                <?php endforeach; ?>
+
             </table>
         </div>
     </div>
